- moved "CellGroup" to its own file
