+++ Tests/AnnoDesigner.Tests/AdjecentCellGrouperTests.cs
@@ -1,5 +1,4 @@
-﻿using System.IO.Abstractions.TestingHelpers;
-using System.Linq;
+﻿using System.Linq;
 using System.Windows;
 using AnnoDesigner.Helper;
 using Xunit;
@@ -32,8 +31,12 @@ namespace AnnoDesigner.Tests
             var postTranspose = Enumerable.Range(0, gridLines.Max(i => i.Length)).Select(i => new bool[gridLines.Length]).ToArray();
 
             for (var i = 0; i < gridLines.Length; i++)
+            {
                 for (var j = 0; j < gridLines[i].Length; j++)
+                {
                     postTranspose[j][i] = preTranspose[i][j];
+                }
+            }
 
             return postTranspose;
         }
